requirements.txt
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==2.2.2
djangorestframework==3.11.0
# fcm-django==0.2.10
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
psycopg2==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.7
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
treq==20.4.1
Twisted==20.3.0
txFCM==0.0.1
urllib3==1.25.9
zope.interface==5.1.0
